requirements.txt
ipython==8.12.2
joblib==1.3.1
langchain==0.0.239
manga109api==0.3.1
matplotlib==3.7.2
numpy==1.23.4
omegaconf==2.3.0
openai==0.27.8
opencv-python-headless==4.8.0.74
pandas==2.0.3
Pillow==9.5.0
pyclustering==0.10.1.2
pydantic==1.10.11
pytorch-metric-learning==2.3.0
pytz==2023.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
tiktoken==0.4.0
torch==2.0.1
torchaudio==2.0.2
torchnet==0.0.4
torchvision==0.15.2
tqdm==4.65.0
